fix: ignore unexported fields on struct extraction
We should ignore unexported fields when extracting
field values from struct.
Because we treat field name with lower case and look up
"yaml" struct field tag later, unexported field name might
conflict with such fields, which produces unexpected results.
This is not a problem if the user is able to change
the name of unexported fields, but sometimes it's hard to
do so. For example, if the struct is generated by a tool,
it would be difficult to change the name of generated
unexported field.